feat: Add getCommandIds function to fetch registered global application commands
This modification adds a new function `getCommandIds` to the `Util.ts` file in the `src/utils` directory. This function is responsible for fetching the registered global application commands from the Discord Client and returning an object containing the command names as keys and their corresponding IDs as values. If there are no commands or an error occurs, an empty object is returned.